+++ integration/setup/setup.go
@@ -12,11 +12,14 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+// Package setup provides integration tests setup helpers.
 package setup
 
 import (
 	"context"
+	"flag"
 	"fmt"
+	"path/filepath"
 	"runtime/trace"
 	"strings"
 	"testing"
@@ -31,6 +34,36 @@ import (
 
 	"github.com/FerretDB/FerretDB/integration/shareddata"
 	"github.com/FerretDB/FerretDB/internal/util/testutil"
+)
+
+// Flags.
+var (
+	targetURLF     = flag.String("target-url", "", "target system's URL; if empty, in-process FerretDB is used")
+	targetBackendF = flag.String("target-backend", "", "target system's backend: '%s'"+strings.Join(allBackends, "', '"))
+
+	targetProxyAddrF  = flag.String("target-proxy-addr", "", "in-process FerretDB: use given proxy")
+	targetTLSF        = flag.Bool("target-tls", false, "in-process FerretDB: use TLS")
+	targetUnixSocketF = flag.Bool("target-unix-socket", false, "in-process FerretDB: use Unix socket")
+
+	postgreSQLURLF = flag.String("postgresql-url", "", "in-process FerretDB: PostgreSQL URL for 'pg' handler.")
+	tigrisURLSF    = flag.String("tigris-urls", "", "in-process FerretDB: Tigris URLs for 'tigris' handler (comma separated)")
+
+	compatURLF = flag.String("compat-url", "", "compat system's (MongoDB) URL for compatibility tests; if empty, they are skipped")
+
+	// Disable noisy setup logs by default.
+	debugSetupF = flag.Bool("debug-setup", false, "enable debug logs for tests setup")
+	logLevelF   = zap.LevelFlag("log-level", zap.DebugLevel, "log level for tests")
+
+	recordsDirF = flag.String("records-dir", "", "directory for record files")
+
+	disablePushdownF = flag.Bool("disable-pushdown", false, "disable query pushdown")
+)
+
+// Other globals.
+var (
+	allBackends = []string{"ferretdb-pg", "ferretdb-tigris", "mongodb"}
+
+	CertsRoot = filepath.Join("..", "build", "certs") // relative to `integration` directory
 )
 
 // SetupOpts represents setup options.
