+++ integration/shareddata/scalars.go
@@ -32,7 +32,7 @@ const (
 // This shared data set is frozen. If you need more values, add them in the test itself.
 var Scalars = &Values[string]{
 	name:     "Scalars",
-	backends: []string{"pg"},
+	handlers: []string{"pg"},
 	data: map[string]any{
 		"double":                    42.13,
 		"double-whole":              42.0,
@@ -115,7 +115,7 @@ var Scalars = &Values[string]{
 // Doubles contains double values for tests.
 var Doubles = &Values[string]{
 	name:     "Doubles",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": tigrisSchema(`"type": "number"`),
@@ -148,7 +148,7 @@ var Doubles = &Values[string]{
 // in diff tests https://github.com/FerretDB/dance.
 var BigDoubles = &Values[string]{
 	name:     "BigDoubles",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": tigrisSchema(`"type": "number"`),
@@ -164,7 +164,7 @@ var BigDoubles = &Values[string]{
 // Tigris JSON schema validator contains extra properties to make it suitable for more tests.
 var Strings = &Values[string]{
 	name:     "Strings",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": `{
@@ -191,7 +191,7 @@ var Strings = &Values[string]{
 // Binaries contains binary values for tests.
 var Binaries = &Values[string]{
 	name:     "Binaries",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": tigrisSchema(`"type": "object", "properties": {"$b": {"type": "string", "format": "byte"}, "s": {"type": "integer", "format": "int32"}}`),
@@ -207,7 +207,7 @@ var Binaries = &Values[string]{
 // ObjectIDs contains ObjectID values for tests.
 var ObjectIDs = &Values[string]{
 	name:     "ObjectIDs",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": tigrisSchema(`"type": "string", "format": "byte"`),
@@ -223,7 +223,7 @@ var ObjectIDs = &Values[string]{
 // Bools contains bool values for tests.
 var Bools = &Values[string]{
 	name:     "Bools",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": `{
@@ -247,7 +247,7 @@ var Bools = &Values[string]{
 // DateTimes contains datetime values for tests.
 var DateTimes = &Values[string]{
 	name:     "DateTimes",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": tigrisSchema(`"type": "string", "format": "date-time"`),
@@ -265,7 +265,7 @@ var DateTimes = &Values[string]{
 // Nulls contains null value for tests.
 var Nulls = &Values[string]{
 	name:     "Nulls",
-	backends: []string{"pg"}, // Not compatible with Tigris as it needs a data type to be set.
+	handlers: []string{"pg"}, // Not compatible with Tigris as it needs a data type to be set.
 	data: map[string]any{
 		"null": nil,
 	},
@@ -274,7 +274,7 @@ var Nulls = &Values[string]{
 // Regexes contains regex values for tests.
 var Regexes = &Values[string]{
 	name:     "Regexes",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": tigrisSchema(`"type": "object", "properties": {"$r": {"type": "string"}, "o": {"type": "string"}}`),
@@ -290,7 +290,7 @@ var Regexes = &Values[string]{
 // Int32s contains int32 values for tests.
 var Int32s = &Values[string]{
 	name:     "Int32s",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": tigrisSchema(`"type": "integer", "format": "int32"`),
@@ -311,7 +311,7 @@ var Int32s = &Values[string]{
 // Timestamps contains timestamp values for tests.
 var Timestamps = &Values[string]{
 	name:     "Timestamps",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": tigrisSchema(`"type": "object", "properties": {"$t": {"type": "string"}}`),
@@ -327,7 +327,7 @@ var Timestamps = &Values[string]{
 // Int64s contains int64 values for tests.
 var Int64s = &Values[string]{
 	name:     "Int64s",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	validators: map[string]map[string]any{
 		"tigris": {
 			"$tigrisSchemaString": tigrisSchema(`"type": "integer", "format": "int64"`),
@@ -349,7 +349,7 @@ var Int64s = &Values[string]{
 // Unsets contains unset value for tests.
 var Unsets = &Values[string]{
 	name:     "Unsets",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	data: map[string]any{
 		"unset": unset,
 	},
@@ -358,7 +358,7 @@ var Unsets = &Values[string]{
 // ObjectIDKeys contains documents with ObjectID keys for tests.
 var ObjectIDKeys = &Values[primitive.ObjectID]{
 	name:     "ObjectIDKeys",
-	backends: []string{"pg", "tigris"},
+	handlers: []string{"pg", "tigris"},
 	data: map[primitive.ObjectID]any{
 		{0x00, 0x01, 0x02, 0x03, 0x04, 0x05, 0x06, 0x07, 0x08, 0x09, 0x10, 0x11}: "objectid",
 		primitive.NilObjectID: "objectid-empty",
